requirements_training.txt
# =============================================================================
# REQUISITOS PARA ENTRENAMIENTO MOBILENET V2
# =============================================================================

# TensorFlow y Keras (versiones recientes)
tensorflow>=2.16.0
keras>=2.16.0

# Procesamiento de imágenes
Pillow>=9.5.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Análisis de datos y visualización
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Google Cloud Storage
google-cloud-storage>=2.10.0
google-auth>=2.20.0

# Utilidades
tqdm>=4.65.0
pathlib2>=2.3.7
python-dotenv>=1.0.0

# Monitoreo y logging
tensorboard>=2.16.0

# Procesamiento paralelo
joblib>=1.3.0

# =============================================================================
# INSTALACIÓN RECOMENDADA:
# =============================================================================
# source venv/bom/activate; pip install -r requirements_training.txt
#
# O para desarrollo:
# pip install -r requirements_training.txt --upgrade
#
# Para GPU NVIDIA:
# pip install tensorflow[gpu]
# =============================================================================
